Separated debounce hook into own file, comments
